climb: store pre and post measures in each log
In each session, allow the user to store a value with its best
measurement after warm-up and after finishing the workout.

This is later used to calculate the force loss on the log detail.

Currently added as POC, with one value per hand.
This should be more customizable, maybe with an option to register extra
values in the log.